Resolved runtime errors in original ST-GCN
Original ST-GCN uses a Conv2D layer for temporal accumulation of
the Gamma sliding window over the incoming frames: to be equivalent
to the author's formulae, the convolution must be done with a fixed
kernel and without the bias term. It uses a strided Conv2D, which
pools the original video sequence instead of mapping it one-to-one:
this is incorrect because classification is then made on the whole
video and not on a frame-by-frame basis.
Replacing striding with the conceptually intended dilation achieves
the purpose - enlarging the temporal receptive field in the
4th and 7th layers.
Original ST-GCN flattens input tensors before the dot product with
the adjacency matrix: this needs validation because it would result
in a linear combination of elements belonging to different batches,
partitions and channels (not the case in the broadcasted
multidimensional MMUL, used in our proposed model).
Original ST-GCN also unnecessarily adds the bias term when Conv2D
is followed by a BatchNorm layer.

+: Rolled out single .pt file checkpointing.
+: Sending email notifications with epoch statistics after each
   epoch.
+: Flushing log file buffers to view logs during program execution.
+: Monotonic learning rate decay by a constant factor.
+: Replaced original ST-GCN's strided Conv2D with a dilated Conv2D
   for temporal accumulation of feature vectors.
+: Rearranged location of common modules.

TODO: Extend our ST-GCN model input with the multiple skeleton
      dimension to process all available skeletons in the frame
      in parallel (another broadcasted dimension, similar to batch).
TODO: Make a branch to test performance difference in temporal
      accumulation of updated graph feature vectors if exchanging
      MMUL with a lower-triangular matrix for convolution with a
      fixed-kernel of ones.
TODO: Validate in original ST-GCN that .view() on the input followed
      by the dot product, inside `tgcn.py`, produces an operation
      equivalent to the principle of work proposed in the author's
      paper.
TODO: Connect to Weights & Biases for interactive model monitoring.