Batching PKU-MMD trials exceeds local memory
Zero padding and masking batch data to construct a multi-dimensional
single tensor of differently sized trials by padding all trials to
the length of the longest recording in the set exceeds local memory.

+: Updated dataset from directory class (for PKU-MMDv2) to batch
   differently sized trials into one tensor by zero padding and
   masking to the length of the longest trial in the set.
+: Corrected MSE loss calculation to be applied elementwise temporally.
+: Weighted CE loss to adjust for class imbalance of the provided
   dataset.

-: Removed iterating over dataloader to test overfitting of the model
   to a single batch.
-: Uncommented logging and emailing blocks.

Note: Model able to overfit on a single sample -> validates that
      it is able to learn.

TODO: Run a debug job on the HPC to validate the dataset batching code.