Supplied time series as batches to original
Using .unfold() to create a view of an underlying original tensor
with slices of the receptive field size, apart by 1 time frame
(for overlapping input window in 1-to-1 segmentation) or by the
number of frames equal to the receptive field size (for
non-overlapping input window in reduced temporal resolution
segmentation).
Now, the non-equal length dataset, applied to the true original
model can be parallelized across multiple GPUs and avoid exceeding
available GPU memory.

+: Unfolding time-series into a batch of overlapping windows for
   parallelizing non-equal trials training across multiple GPUs.

-: String splitting in model loading from a checkpoint.

TODO: Use the technique to also parallelize realtime and adapted
      models.